Fix buffer wraparound for regex searching
SearchNext/SearchPrevious actions doesn't work in almost cases. The
actions try to search between head point of current line on vi mode
cursor to tail point of a line of one level above the line. For example,
when a start point is at (3, 0) and an end point position is determined
as (2, 119) on a grid(80x120) for a searching, the searching doesn't
find anything.

The reason why is that the start point has already been beyond the end
point(`(2, 119) <= (3, 0)`). So It is recognized that the search
traversing has already been completed and it exits immediately on
starting.

So consider whether needs to wrap around to opposite side of scroll
buffer at an scroll buffer side or not, in addition to searching start
point and the end one, to detect search traversing completion.