Fix ScrollHalfPageUp action's regression
The regression is that cursor doesn't move to topmost on scrollback
buffer and keep it near bottommost when invokes ScrollHalfPageUp action
of vi mode even if repeats it. Also it is happen even if vi mode cursor
is on above bottommost when in topmost viewport. So looks like that the
cursor moves to bottommost line on topmost viewport even though try to
move cursor to upper line.

This regression is came from 3bd5ac2 which introduces unified index
type. The reason why is that old index type for viewport minimum index
is zero, but new index type minimum one is zero or minus value. So vi
mode cursor scrolling for new index type need to consider the new rule.
So use `Boundary::Grid`, which is range of entire grid, as cursor point
clamping instead of `Boundary::Cursor`, which is range of viewport.