Fix regex search regression
The regression is that vi cursor doesn't jump to match in almost cases
when invokes SearchNext/SearchPrevious actions. (The exception is that
vi cursor is on topmost line of scrollback buffer or on bottompost and
rightmost point of one.)

The original problem of this regression is that the regex searching
skips one character follows wide character without wide character spacer
unintentionally.