Respond to XTVERSION query
XTVERSION requests take the form of CSI > [0] q, and
are responded to with a freeform string. Differentiate
between XTVERSION and DECSCUSR (CSI n SP q), and reject
any XTVERSION that specifies a Ps other than 0 (admit
an absent Ps as 0, per the spec). Respond to XTVERSION
with Alacritty and the Cargo package version of
alacritty_terminal (it's this core that's relevant for
determining capabilities).

This is necessary for Notcurses to unambiguously identify
Alacritty (i.e. as opposed to relying on TERM), and is
implemented by many other terminal emulators.