Basic touch support: tapping can now be used in lieu of clicking.
When a touch event is ended (by the finger being lifted up), two mouse
events are sent to emulate a click. All other touch events (Down, Move,
Cancel) are simply ignored. This makes no decisions about how other
touch events should be treated. However, a touch, followed by a long
move, followed by the end of the touch, is still treated as a click.

This change ignores the message bar, so bugs likely lurk there.

Probably because of winit issue #1996 [1], the click is received in the wrong
location unless there was an intervening move event (on Wayland). This
bug will be easy to work around once Down and Move touches are no longer
ignored.

[1] https://github.com/rust-windowing/winit/issues/1996