Extended touch support: keep track of touch position
The new TouchState object will track information in parallel to Mouse.
Right now, only the position of the last touch is tracked. As a
by-product of this tracking, this commit contains a workaround for winit
issue 1996 with no extra effort (the position associated to
TouchPhase::Ended is now simply ignored, in favor of the known-good last
position).

Multitouch is entirely ignored in the code, as TouchState stores only a
single position.