Add unix socket for spawning new windows
This adds an IPC socket on unix platforms which make it possible to
communicate with existing Alacritty instances using the `alacritty msg`
subcommand. The socket is determined using the `ALACRITTY_SOCKET`
environment variable, which should make it work automatically when used
from the target terminal.

If the environment variable is not set, a socket is determined at random
based on the first socket file found in /tmp. For the existing
functionality (creating new windows), this should work out great. Though
it might not be ideal with other usecases for an IPC socket.