Touch support: detecting drags and ignoring them
(Previously, a drag was treated as a click on the ending position.)

The constant DRAG_MINIMUM_DISTANCE2 determines how far a finger must
travel before the event is a drag, rather than a click. At the moment,
this is set slightly large, to ensure that clicks are never missed. As
there's no drag functionality, this is a conservative choice.