Fix timer scheduling
This resolves two issues with the timer scheduling.

The primary bug resolved by this patch is related to the order of
scheduled events. The scheduler always assumes that all events in the
queue are sorted, however the `get_mut` method made it possible to
mutate a timer's deadline allowing the order of events to get messed up.
Since mutating in-place would still require re-insertion into the queue
anyway, this method has been removed in favor of using `unschedule` and
`schedule` to re-schedule an event.

The other issue is due to repeating timers sticking around permanently
with the multi-window patch, so to ensure we don't leak memory or cause
the scheduler to get extremely slow, all scheduled events for a window
are now removed after the window is closed.