Move dpr calculations into its own function, removing duplication.
This also applies the X11 sanity check every time the dpr is updated.
Previously, the check for suspicious dpr values wasn't carried out when
the dpr was updated - for example changing between monitors. Now it uses
the same code path as elsewhere so the check is carried out.

This commit also adds code to update the dpr when the config is
reloaded.