Add font-size presets
This patch adds a "font.preset_size" item to the configuration file.
This item holds an array of preset font sizes which can be selected with
the new FontSize1, FontSize2, and FontSize3 commands.  Bindings for
these commands default to Ctrl(/Cmd)-1/2/3.

At the moment, this is only a proof of concept to demonstrate a feature
I would use.  Ideally, an implementation of this feature would place the
preset sizes directly in the bindings, eliminating the extra array.
This would require a parameterized action, but it was not obvious how to
set one up (if it is currently possible at all), hence this sub-ideal
approach.