Fix crash when clears the viewport and the vi cursor on scrollback
Alacritty currently crashes if the cursor is outside of the upper
display area when rendering contents. The situation is caused when
clears the viewport and the vi cursor is on the scrollback buffer. The
easy reproduction steps are below. They are from
https://github.com/alacritty/alacritty/issues/5544, but with a few
fixes.

```sh
# Create a number of scrollback buffers.
$ seq 100

# Delay clearing long enough to enter vi mode and jump to top of
# scrollback
$ sleep 5; clear
# ctrl+space, g
# wait for clear to execute
```

Maybe we should solve the crash, but we can also fix it by fitting the
vi cursor into the display area. It generally makes sense regardless of
this problem.