Remove trailing newline from strings taken from hints or some selections
Currently hint regex doesn't match to newline character because
Alacritty treats wrapped lines as a single line and end of not wrapped
line is end of match. So strings taken from hints should not contain
trailing newline character at the end of it.

This behavior should also be consistent for all selections except lines
one due to no way to contain a newline character at the end of all
selection.

Doing them leads to a consistent behavior for all hint actions.

Fixes #5697.