Fix text joining by block selection contained last column
This problem is that copied lines are joined when block selection
contains the last column.

Assume a terminal grid like below.

```
aaa
bbb
```

When we select whole lines by block selection and copy it, expected
copied text is `aaa\nbbb`, but actual one is `aaabbb`.

This is introduced by 8e584099.