Fix macOS release CI
Since the CI machine is amd64, it is not possible to just execute the
aarch64 binary to test the application. So instead of running `cargo
test`, we just use `cargo build` instead.

Building on the non-release CI instead of just checking also allows us
to make sure linking works properly before a release fails due to it.