Optimize terminal full damage
When terminal should be fully damaged we only want to occupied cells,
thus now alacritty_terminal always returns `TermDamageIterator`, but
when terminal was fully damaged it damages all the lines from 0 up to
line.occ. That helps a lot with something like `yes` or when scrolling
scrollback buffer by reducing the amount of damaged area.