Show the history line indicator also when scrolled
There is currently no scrollbar, but also not any kind of visual
indication when the user scrolled the buffer into the history.

There is however the numerical line indicator that is drawn when in
search mode or in vi cursor mode.
This changes also draws the indicator any time where the current
display offset is non-zero.

Related to #775