Allow overlapping graphics.
If a graphic is added over another one, the implementation now checks if new
graphic has transparent pixels in every cell. If so, the graphic is appended to
the cell, instead of replacing the previous one.

SmallVec is used to prevent heap allocation when the cell only contains a single
graphic. This should be the most common scenario.

The GPU will store up to 100 textures. If another texture is added when there
are 100, the oldest one is deleted.