Optimize graphics replacement.
A simple optimization for inserting graphics is to detect when a new graphic is
replacing completely an existing one. If both graphics have the same size, and
the new one is opaque, we can assume that the previous graphic will not be
displayed anymore, so it is not considered when update the graphics list in a
single cell.

This commit also adds serde implementation for GraphicsCell. This was used to
debug the new logic.