Only set dirty when cursor_hidden is changed in update_cursor_blinking
This little change reduces the latency of 1 frame when typing in the
terminal. The reason is that it removes an extra redraw before receiving
the visual update from the PTY, and that allows not having to wait until
a frame can be rendered again.

Also, simplify the logic in on_typing_start by directly calling
update_cursor_blinking instead of copying its logic.