Remove self.dirty from update_highlighted_hints condition
Separate the fact that the hint highlighting is dirty from the fact that
the terminal must be redrawn, and mark the highlighted hint as dirty
when the layout changes.

Although they add a few lines of code, these changes make a bit easier
to follow what is going on with the hint highlighting by being more
explicit.