Stop marking the terminal damaged after rendering
With commit ee40518dcc9a7eef38048c2b95faab75b4d50b72, damaging the
terminal now signifies that it needs to be rendered as soon as possible.

Use the next_frame_damage_rects field instead of damaging the terminal
after rendering when we want the next frame to start with damage, as it
already exists for this purpose.