[docs] use curl instead of local files if possible
Some commands could be done using the in-memory temporary files instead of requiring users to clone the repo.  This way alacritty can be installed with `cargo install`, and the terminfo can be updated without any `git clone.  Although it might be worth while to set up a separate installation section for these tricks?