create new windows with Cmd+N by default
This changes the default Cmd+N binding on macOS to create a new window
rather than spawning a new instance.

Initially this change was held back for further testing of the
multi-window feature. At this point all significant issues found with it
have been fixed so it should be ready for prime-time now.