Add EmptySelection key binding mode
With this, a bind can be configured to trigger based on the presence or
absence of a selection.

For example, this config emulates windows terminal style copy paste:
- { mouse: Right,  mode: ~EmptySelection, mods: Shift, action: Copy }
- { mouse: Right,  mode: ~EmptySelection, mods: Shift, action: ClearSelection }
- { mouse: Right,  mode: EmptySelection,  mods: Shift, action: PasteSelection }

In the above example, when no text is selected shift-rightclick will
paste. If text is selected, then shift-rightclick will copy and clear
the selection.