Mention all the actions in the man page
A lot of actions were left without a notice, so the only way to figure
out some of them was to browse source code. An example of such
actions were `Maximize` and `Minimize`, since we don't have a binding
for them, so they were left even in alacritty-bindings(5).

Explicitly list all the bindings we have. The search bindings were
also not accuratelly restricted, since we allow them to be executed
inside regardless of mode.