Start terminal even if `working_directory` is invalid, show a warning
Add a way to communicate errors from pre_exec() to the main process, use
it for errors to change directory and to set up the session.

Remove the die! macro, all errors are handled in a better way now.

Only MacOS and UNIX are affected, Windows is using different code.