Fix ime position
In the past, only the bottom left corner point was passed to ime, so the
position was `point.line + 1` to account for current line. However,
a rectangle is now passed, and the coordinate passed should be the top
left corner of the rectangle. The current line height should not be
added in this case.