linux: Don't fail exec()s when shell's pwd doesn't exist
Currently, exec()s (e.g. for hint executables) fail when the shell's
working directory doesn't exist with a fairly inscrutable:

    [26239.852699733s] [WARN ] [alacritty] Unable to launch xdg-open with args ["https://tycho.pizza"]

It turns out that the stdlib will fail these executions if it can't set
the working directory. Since Linux appends the string " (deleted)" during a
readlink("/proc/pid/cwd") for a directory that has been deleted, this
directory won't exist, so this will always fail.

Let's only set the current working directory if it actually exists.

...it may be worth only doing this in the hints case, but especially then
it doesn't seem super relevant what (if any) directory the shell's process
is in for hints.

Alternatively, we could fall back to chdir("/proc/pid/cwd") directly, but
that seems like overkill to me (and won't render the same as the original
for people that do getcwd() later).

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>