Fix selection clearing in kitty keyboard mode
When Kitty's keyboard protocol is used and Report all keys as escape
codes flag (8) is enabled, modifier key escape codes trigger the usual
"write something to the terminal" code path, which clears the selection
/ scrolls down etc.

This behavior is mostly unexpected, and makes some actions more painful
to perform (for instance copying text becomes harder: hitting CTRL to
begin the CTRL+SHIFT+C sequence clears the selection).

This patch clears the selection only if the key event is not a modifier
key, which aligns with Alacritty's usual behavior.

Fixes #8509.