Remove unused `stroke` attributes from SVG logos
This prevents warnings in Qt apps using the QtSvg library when displaying
the affected SVG logos. The warning:

qt.svg: <input>:LINE:COLUMN: Could not resolve property: #linearGradient10962

On Linux, for example, you'll see that warning by opening any of the
affected SVGs with Gwenview[1] from the command line. You can also
occasionally see it in the system journal when KDE Plasma loads the
Alacritty logo for the first time to display it on the taskbar.

That `#linearGradient10962` property does not exist in any of the SVGs,
so this removes all references to it in `stroke` attributes.

This also removes other unused `stroke-*` attributes that do not affect
the SVGs visually.

[1] https://apps.kde.org/gwenview/