Fix file not found errors when editing alacritty.toml
Alacritty monitors its config file and reloads as soon as any changes
happen. This leads of "File not Found" errors being printed pretty much
ever time I edit my config file.

This turned out to be because on save vim moves the file, writes the
changes and then moves it back. Aalacritty just happens to always
try and write to the file while it's not there.

This commit Adds special handling for renames and deletions. If either
operation is currently the most recent event we discard the whole
event queue with the assumption the file does not currently exist.