Ignore when config file is moved
As this is practically the same as deleting the file it should be
treated the same and ignored.

Moving a file from somewhere to the config file is not ignored and the config
will reload as normal.

This fixes file not found errors when saving the config file in some
editors. This doesn't consider the possible race condition where a modify and
rename event both exist in the queue.